+++ changelog.md
@@ -148,3 +148,5 @@ V1.17.0
 V1.18.0
 
 * Add ability to change attachment position of entities
+* Add ability to attach entities to other entities
+* Add vehicle name to pre-existing add overlay